Legger til felt for virksomeht. Endrer rekkefølge og oppsett i PDF. Flytter Virksomhet og tilhørende klasser til egen fil for bedre lesbarhet og oversikt
